+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.2.0
 - The parse method is now on the top-level namespace (`Wifimap.parse`).
 - Added `Wifimap.parse_file(filename)`.
+- Added support for [sniff-probes](https://github.com/brannondorsey/sniff-probes).